% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExportsToDate.R
\name{getExportsToDate}
\alias{getExportsToDate}
\title{Get exports to date}
\usage{
getExportsToDate(countryName)
}
\arguments{
\item{countryName}{Enter a specific country to search for; case sensitive.}
}
\value{
A summary table to the console.
}
\description{
Just returns a table of cumulative exports to date over the last few years
}
