% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuturePrices.R
\name{getFuturePrices}
\alias{getFuturePrices}
\title{Get futures prices}
\usage{
getFuturePrices(writeCSV = TRUE, env = .GlobalEnv)
}
\arguments{
\item{writeCSV}{writes any new data to the local data folder. TRUE by default.}
}
\description{
This function queries Quandl for the second month Arabica/Robusta contracts and updates a local csv file: futurePrices.csv. It will automatically go back far enough to get any missing values from the local file.
NB Quandl API key is set in the environmental variables
}
