% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFuturesCurve.R
\name{getFuturesCurve}
\alias{getFuturesCurve}
\title{Get coffee futures curve}
\usage{
getFuturesCurve()
}
\value{
A plotly graph to the Viewer. That's it.
}
\description{
Runs a graph for Arabica & Robusta futures curves, based on yesterday's settle price.
}
