% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICOhistorical.R
\name{getICOhistorical}
\alias{getICOhistorical}
\title{Get ICO historical data}
\usage{
getICOhistorical(env = .GlobalEnv)
}
\value{
icoData A dataframe containing data back to 1990.
}
\description{
Downloads data from the excel files on the ICO historical data website, for production, domestic consumption, crop year exports and opening stocks. Writes a csv file to the coffeestats data folder and returns a dataframe called icoData to the global environment.
}
