% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICOprices.R
\name{getICOprices}
\alias{getICOprices}
\title{Get ICO  prices.}
\usage{
getICOprices(writeCSV = TRUE)
}
\description{
Gets daily prices for the ICO composite and 4 groups from the website. Downloads the Excel file for the last two months, and if needs to then goes through the pdfs. Saves in a file called icoPrices.csv in coffeestats/data. Also calls loadICOprices() which returns a dataframe called prices.
If writeCSV = FALSE, it just returns the dataframe without updating the csv (I can't remember why you would want to do that, so TRUE by default).
}
\examples{
getIcoPdfPrices()
}
