% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getICOproduction.R
\name{getICOproduction}
\alias{getICOproduction}
\title{Get ICO production/consumption figures}
\usage{
getICOproduction()
}
\value{
Returns a dataframe called icoFlow to the global environment, and a file with today's date to the data folder
}
\description{
\code{getICOproduction()} scrapes the ICO production/consumption pdf tables from the website and saves them into a local csv file
}
