% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMTS.R
\name{getMTS}
\alias{getMTS}
\title{Get Monthly Trade Statistics}
\usage{
getMTS(refresh = TRUE, writeFile = TRUE)
}
\arguments{
\item{refresh}{Do you want to download new data from the ICO website?}

\item{writeFile}{Do you want to save this to the local data file.}
}
\value{
Returns a dataframe called mts to the global environment.
}
\description{
\code{getMTS()} scrapes the ICO website for export data and saves the last six months to a local csv file.
}
