% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProduction.R
\name{getProduction}
\alias{getProduction}
\title{Import production data from an input file stored on my computer.}
\usage{
getProduction(env = .GlobalEnv)
}
\description{
Doesn't take any arguments, and returns a dataframe called prodAll to the global environment.
}
\examples{
getProduction()
}
