% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUSDA.R
\name{getUSDA}
\alias{getUSDA}
\title{Get USDA data}
\usage{

  getUSDA(url = "https://apps.fas.usda.gov/psdonline/downloads/psd_coffee_csv.zip")
}
\arguments{
\item{url}{Should be a link to the most recent USDA coffee \strong{zip file}.}
}
\value{
Also returns a dataframe called usda, which is loaded from the usda-tidy.csv using loadUSDA() under the hood.
}
\description{
Downloads a zip file containing a csv from the USDA website, imports it, cleans it and write two new csvs: USDA-tidy.csv containing a tidy dataframe of the data, and USDA-flowsheets.csv containing flowsheet info for the last five years, for importing into Excel.
}
\examples{
getUSDA()
}
