% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goCompare.R
\name{goCompare}
\alias{goCompare}
\title{Go Compare}
\usage{
goCompare(countryName)
}
\arguments{
\item{countryName}{Case sensitive, needs to be a country in the system.}
}
\description{
Runs a comparison for a specific country of various production estimates, and returns graphs of exports so far and production.
}
