% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCountryList.R
\name{loadCountryList}
\alias{loadCountryList}
\title{loadCountryList}
\usage{
loadCountryList(env = .GlobalEnv)
}
\value{
countries
}
\description{
Just imports the list of countries, continents, and importer/exporter type.
}
