% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFlowsheetData.R
\name{loadFlowsheetData}
\alias{loadFlowsheetData}
\title{loadFlowsheetData}
\usage{
loadFlowsheetData(env = .GlobalEnv)
}
\value{
flow
}
\description{
Import summary data from my-flowsheets into a dataframe called flow.
}
