% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadICOhistorical.R
\name{loadICOhistorical}
\alias{loadICOhistorical}
\title{Load ICO historical data}
\usage{
loadICOhistorical(env = .GlobalEnv)
}
\value{
Dataframe called icoData to the global environment
}
\description{
Just loads ico historical data from a local file, which should be in the coffeestats data folder, and have the form YYYY-MM-DD-ico-historical
}
