% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadICOprices.R
\name{loadICOprices}
\alias{loadICOprices}
\title{Load ICO prices}
\usage{
loadICOprices()
}
\value{
Dataframe called icoPrices to the global environment
}
\description{
Just loads data from a local file called icoPrices.csv, which should be in the coffeestats data folder and returns a dataframe called prices.
}
