% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadICOproduction.R
\name{loadICOproduction}
\alias{loadICOproduction}
\title{Loads ICO production/consumption}
\usage{
loadICOproduction(env = .GlobalEnv)
}
\value{
Dataframe called icoFlow to the global environment
}
\description{
Just loads  data from a local file, which should be in the coffeestats data folder, and have the form "2017-01-01-icoFlow.csv".
}
