% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMTS.R
\name{loadMTS}
\alias{loadMTS}
\title{Load monthly trade statistics}
\usage{
loadMTS(env = .GlobalEnv)
}
\value{
Dataframe called mts to the global environment
}
\description{
Just loads export data from a local file, which should be in the coffeestats data folder, and called export-data.csv.
}
