% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadScaleExports.R
\name{loadScaleExports}
\alias{loadScaleExports}
\title{Load scale exports file.}
\usage{
loadScaleExports(env = .GlobalEnv)
}
\value{
Dataframes called scaleExports and scaleExportsRaw to the global environment
}
\description{
Just loads the scaleExports and scaleExportsRaw csv files from a local directory. Doesn't do anything else.
}
