% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUSDA.R
\name{loadUSDA}
\alias{loadUSDA}
\title{Load USDA files}
\usage{
loadUSDA(env = .GlobalEnv)
}
\value{
Dataframe called usda to the global environment
}
\description{
Just loads usda data from a local file, which should be in the coffeestats data folder, and have the form YYYY-MM-DD-usda-tidy.csv.
}
