% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predExports.R
\name{predExports}
\alias{predExports}
\title{Exports prediction function}
\usage{
predExports(getCountry, table = TRUE)
}
\arguments{
\item{getCountry}{Enter a specific country to search for; case sensitive.}

\item{table}{Returns a summary table to the console if true.}
}
\value{
A plotly graph and a table.
}
\description{
Allows you to enter a specific country and explore the scaleExports/MTS data. Useful for flowsheet predictions etc.
}
