% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDataDir.R
\name{setDataDir}
\alias{setDataDir}
\title{Set the data folder.}
\usage{
setDataDir(folderLocation = "~/Dropbox/Work/coffeestats/data",
  env = .GlobalEnv)
}
\description{
I use this to store data in a dropbox folder. All other functions should relate back to this function, so you only need to change it once if the folder moves.
Almost definitely not best practice, but just stores the folder location as a string in the global environment. It seems to work for me
}
\examples{
setDataDir(folderLocation = "~/Dropbox/Work/coffeestats")
}
