% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProduction.R
\name{getProduction}
\alias{getProduction}
\title{Import production data from an input file stored on my computer}
\usage{
getProduction()
}
\description{
Import production data from an input file stored on my computer
}
\examples{
getProduction()
}
