% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScaleExports.R
\name{getScaleExports}
\alias{getScaleExports}
\title{Title}
\usage{
getScaleExports(refresh = FALSE, render = FALSE)
}
\arguments{
\item{refresh}{Set as TRUE to update monthly trade data from ICO website.}

\item{render}{Set as TRUE to run a R Markdown report on the data.}
}
\value{
Calls loadScaleExports to return scaleExports and scaleExportsRaw
}
\description{
Title
}
\examples{

}
